# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List tourist programs
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    patch:
      operationId: updateProgram
      tags: [programs]
      summary: Update a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /program-bookings:
    get:
      operationId: listMyProgramBookings
      tags: [program-bookings]
      summary: List the authenticated customer's program booking requests
      responses:
        "200":
          description: The customer's booking requests, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgramBooking"
    post:
      operationId: createProgramBooking
      tags: [program-bookings]
      summary: Submit a program booking request (auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramBookingInput"
      responses:
        "201":
          description: Booking request created with an official TRV reference number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramBooking"
  /program-bookings/{id}:
    get:
      operationId: getProgramBooking
      tags: [program-bookings]
      summary: Get one booking request (owner or staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramBooking"
  /program-bookings/{id}/cancel:
    post:
      operationId: cancelProgramBooking
      tags: [program-bookings]
      summary: Customer cancels their own booking request (only before confirmation)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramBooking"
  /admin/program-bookings:
    get:
      operationId: listProgramBookingsAdmin
      tags: [program-bookings]
      summary: List all program booking requests (staff with bookings permission)
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProgramBookingStatus"
      responses:
        "200":
          description: All booking requests, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgramBooking"
  /admin/program-bookings/{id}/status:
    patch:
      operationId: updateProgramBookingStatus
      tags: [program-bookings]
      summary: Update a booking request status (staff); notifies the customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramBookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramBooking"
...
components:
  schemas:
...
    DailyItineraryItem:
      type: object
      required: [day, titleAr, titleEn, descriptionAr, descriptionEn]
      properties:
        day: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
    ProgramHotel:
      type: object
      required: [nameAr, nameEn, stars, city]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        stars: { type: integer }
        city: { type: string }
    Program:
      type: object
      required: [id, titleAr, titleEn, price, days, imageUrl, status, featured, isActive, createdAt]
      properties:
        id: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: ["string", "null"] }
        bookingTerms: { type: ["string", "null"] }
        cancellationPolicy: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        roomTypes:
          type: array
          items: { type: string }
        meals: { type: ["string", "null"] }
        transport: { type: ["string", "null"] }
        tours:
          type: array
          items: { type: string }
        programDate: { type: ["string", "null"] }
        notesAr: { type: ["string", "null"] }
        notesEn: { type: ["string", "null"] }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProgramInput:
      type: object
      required: [titleAr, titleEn, price, days]
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: string }
        bookingTerms: { type: string }
        cancellationPolicy: { type: string }
        destination: { type: string }
        roomTypes:
          type: array
          items: { type: string }
        meals: { type: string }
        transport: { type: string }
        tours:
          type: array
          items: { type: string }
        programDate: { type: string }
        notesAr: { type: string }
        notesEn: { type: string }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
    ProgramUpdate:
      type: object
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: string }
        bookingTerms: { type: string }
        cancellationPolicy: { type: string }
        destination: { type: string }
        roomTypes:
          type: array
          items: { type: string }
        meals: { type: string }
        transport: { type: string }
        tours:
          type: array
          items: { type: string }
        programDate: { type: string }
        notesAr: { type: string }
        notesEn: { type: string }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
    ProgramBookingStatus:
      type: string
      enum: [draft, submitted, under_review, awaiting_availability, awaiting_payment, paid, confirmed, completed, rejected, cancelled]
    ProgramBookingHistoryEntry:
      type: object
      required: [status, at]
      properties:
        status: { type: string }
        at: { type: string }
        note: { type: ["string", "null"] }
        by: { type: string }
    ProgramBooking:
      type: object
      required: [id, requestNumber, userId, programId, programTitleAr, programTitleEn, programPrice, programCurrency, fullName, email, phone, nationality, nationalityCode, adults, children, infants, travelDate, rooms, status, statusHistory, createdAt, updatedAt]
      properties:
        id: { type: string }
        requestNumber: { type: string }
        userId: { type: string }
        programId: { type: integer }
        programTitleAr: { type: string }
        programTitleEn: { type: string }
        programDestination: { type: ["string", "null"] }
        programPrice: { type: number }
        programCurrency: { type: string }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string }
        nationality: { type: string }
        nationalityCode: { type: string }
        adults: { type: integer }
        children: { type: integer }
        infants: { type: integer }
        travelDate: { type: string }
        returnDate: { type: ["string", "null"] }
        rooms: { type: integer }
        roomType: { type: ["string", "null"] }
        specialRequirements: { type: ["string", "null"] }
        customerNotes: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/ProgramBookingStatus" }
        statusHistory:
          type: array
          items: { $ref: "#/components/schemas/ProgramBookingHistoryEntry" }
        adminNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProgramBookingInput:
      type: object
      required: [programId, fullName, email, phone, nationality, adults, travelDate, rooms]
      properties:
        programId: { type: integer }
        fullName: { type: string, minLength: 3 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 6 }
        nationality:
          type: string
          description: ISO alpha-2 code or canonical country name (resolved server-side against the master list)
        adults: { type: integer, minimum: 1, maximum: 50 }
        children: { type: integer, minimum: 0, maximum: 50 }
        infants: { type: integer, minimum: 0, maximum: 20 }
        travelDate: { type: string, description: "YYYY-MM-DD" }
        returnDate: { type: string }
        rooms: { type: integer, minimum: 1, maximum: 50 }
        roomType: { type: string }
        specialRequirements: { type: string }
        customerNotes: { type: string }
    ProgramBookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/ProgramBookingStatus" }
        note:
          type: string
          description: Optional message to the customer recorded on the timeline and included in the notification
        adminNotes: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT